if condition

+++ python-if_else_loops_functions/1-last_digit.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 import random
 number = random.randint(-10000, 10000)
-number = int(number)
+number = float(number)
 
 last_digit = int(str(number)[-1])
 if last_digit > 5:
